🔨 extract OwidGoogleAuth,  GdocBase
